Implement NYX_OS_INFO_MANUFACTURING_VERSION query
:Release Notes:
The OSInfo module now provides the manufacturing version value found in
WEBOS_DISTRO_MANUFACTURING_VERSION, which is intended to be passed in
on the CMake command line.

:Detailed Notes:
Implemented support for querying the operating system's manufacturing
version. This requires version 5.2.0 of nyx-lib (which defines the
relevant enumeration value).

:Testing Performed:
Compiled locally and tested on desktop and emulator.

:QA Notes:

:Issues Addressed:
[GF-12714] Supply a "manufacturing version" for webOS
[GF-13449] Add implementation of OSInfo query 
           NYX_OS_INFO_MANUFACTURING_VERSION

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>
Change-Id: Ib8cd185f0553c3cc74236e59ec15c73fc27a02d3
Reviewed-on: https://g2g.palm.com/3169
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>