Build modules based on NYX_MODULES_REQUIRED definition
:Release Notes:
Build only the modules specified in NYX_MODULES_REQUIRED string.

:Detailed Notes:
This change is compliant with the other changes in meta layers wherein
 all the modules required for a particular distribution or particular
machine will be included in the NYX_MODULES_REQUIRED string. This change
checks if the NYX_MODULES_REQUIRED string is empty, in which case it builds
all the modules, however if its non-empty it builds only the modules
specified.
Also the other change added is that the "battery" and "charger" modules
are now built independant of "system" module.

:Testing Performed:
Verfied that only the modules specified in NYX_MODULES_REQUIRED string
are built and installed.

:Issues Addressed:
[CTO-188] Modify the CMake Script

Open-webOS-DCO-1.0-Signed-off-by: Sapna Todwal <sapna.todwal@lge.com>

Change-Id: I075efadb4c5567d721332483f18efba8cceba620
Reviewed-on: https://g2g.palm.com/6269
Reviewed-by: DCO Verification
Tested-by: Sapna Todwal <sapna.todwal@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>